Se agregan los controllers

+++ lab_1/src/main/java/com/tbd_grupo_8/lab_1/services/OrdenService.java
@@ -15,7 +15,7 @@ public class OrdenService {
     public List<Orden> findAll() {
         return ordenRepository.findAll();
     }
-    public Orden findById(int id) {
+    public Orden findById(long id) {
         return ordenRepository.findById(id);
     }
     public Orden save(Orden orden) {
